requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
gitdb2==2.0.5
GitPython==2.1.11
idna==2.7
jsonschema==2.6.0
paramiko==2.6.0
pycparser==2.19
pycryptodome==3.8.1
PyNaCl==1.3.0
PyYAML==3.13
requests==2.20.1
six==1.12.0
smmap2==2.0.5
texttable==0.9.1
urllib3==1.24.1
websocket-client==0.56.0
psycopg2-binary==2.8.3
